+++ Node/lib/recognizers/event-recognizer.js
@@ -1,16 +1,18 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
-var ReferralRecognizer = (function () {
-    function ReferralRecognizer(recognizer, options) {
+var EventRecognizer = (function () {
+    function EventRecognizer(recognizer, options) {
         if (options === void 0) { options = {}; }
         this.recognizer = recognizer;
         options.referral = (options.referral !== undefined ? options.referral : true);
         options.postback = (options.postback !== undefined ? options.postback : true);
+        options.optin = (options.optin !== undefined ? options.optin : true);
         options.referralValue = (options.referralValue !== undefined ? options.referralValue : true);
         options.postbackValue = (options.postbackValue !== undefined ? options.postbackValue : true);
+        options.optinValue = (options.optinValue !== undefined ? options.optinValue : true);
         this.options = options;
     }
-    ReferralRecognizer.prototype.recognize = function (context, done) {
+    EventRecognizer.prototype.recognize = function (context, done) {
         var result = { score: 0.0, intent: null };
         if (context.message.source !== 'facebook') {
             done(null, result);
@@ -39,15 +41,34 @@ var ReferralRecognizer = (function () {
             var postback = context.message.sourceEvent.postback;
             var entity = {
                 entity: postback.payload,
-                type: 'referral',
+                type: 'postback',
                 score: 1.0,
-                facebook: postback.payload
+                facebook: postback
             };
             if (this.options.postbackValue) {
                 result.intent = postback.payload.toLowerCase();
             }
             else {
-                result.intent = 'POSTBACK';
+                result.intent = 'postback';
+            }
+            result.score = 1.0;
+            result.entities = [entity];
+            done(null, result);
+            return;
+        }
+        if (context.message.sourceEvent.optin !== undefined && this.options.optin) {
+            var optin = context.message.sourceEvent.optin;
+            var entity = {
+                entity: optin.ref,
+                type: 'optin',
+                score: 1.0,
+                facebook: optin
+            };
+            if (this.options.postbackValue) {
+                result.intent = optin.ref.toLowerCase();
+            }
+            else {
+                result.intent = 'optin';
             }
             result.score = 1.0;
             result.entities = [entity];
@@ -56,6 +77,6 @@ var ReferralRecognizer = (function () {
         }
         done(null, result);
     };
-    return ReferralRecognizer;
+    return EventRecognizer;
 }());
-exports.ReferralRecognizer = ReferralRecognizer;
+exports.EventRecognizer = EventRecognizer;